citycabbie/backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
psycopg2==2.9.9
python-dotenv==1.0.1
redis==5.0.3
sqlparse==0.4.4
timedelta==2020.12.3
typing_extensions==4.10.0
tzdata==2024.1
